requirements.txt
django==4.2.11
djangorestframework==3.16.1
jwt==1.4.0
djangorestframework-simplejwt==5.5.1